+++ test/test_utils.jl
@@ -29,11 +29,11 @@ end
 
     source = RefSource("src", FixedProfile(5), FixedProfile(10), FixedProfile(5),
         dict, [])
-    sink = RefSink( "sink", FixedProfile(20), dict, dict)
+    sink = RefSink( "sink", FixedProfile(20), dict, dict, Array{Data}([]))
     stor = RefStorage("stor", FixedProfile(60), FixedProfile(1), FixedProfile(1),
-        FixedProfile(0), Power, dict, dict, [])
-    sink_em = RefStorageEmissions("sink-em", FixedProfile(1), FixedProfile(1),
-        FixedProfile(1), FixedProfile(1), CO2, dict, dict, [])
+        FixedProfile(0), Power, dict, dict, Array{Data}([]))
+    stor_em = RefStorageEmissions("sink-em", FixedProfile(1), FixedProfile(1),
+        FixedProfile(1), FixedProfile(1), CO2, dict, dict, Array{Data}([]))
     av = GenAvailability("av", array)
 
     get_types(𝒩) = unique(map(n -> typeof(n), 𝒩))
@@ -51,7 +51,8 @@ end
             @test node_types[GenAvailability] == 4
 
             @test node_types[Storage] == 3
-            @test node_types[RefStorageEmissions] == 4
+            @test node_types[RefStorage{ResourceCarrier{Float64}}] == 4
+            @test node_types[RefStorage{ResourceEmit{Float64}}] == 4
         end
     end
 
@@ -65,7 +66,7 @@ end
         if haskey(indexes, NetworkNode)
             @test indexes[NetworkNode] < indexes[Storage]
             @test indexes[NetworkNode] < indexes[Availability]
-            @test indexes[Storage] < indexes[RefStorageEmissions]
+            @test indexes[Storage] < indexes[RefStorage{ResourceCarrier{Float64}}]
             @test indexes[Availability] < indexes[GenAvailability]
         end
     end
@@ -79,7 +80,7 @@ end
     end
 
     @testset "Test 2" begin
-        node_types = EMB.collect_types(get_types([source, sink, stor, sink_em, av]))
+        node_types = EMB.collect_types(get_types([source, sink, stor, stor_em, av]))
 
         test_type_ranking(node_types)
         sorted_node_types = EMB.sort_types(node_types)